requirements.txt
attrs>=19.3
click>=7.0
requests>=2.13.0
